--- a/src/config.zig
+++ b/src/config.zig
@@ -67,6 +67,8 @@ const ConfigProcess = struct {
     clock_epoch_max_ms: u64 = 60000,
     clock_synchronization_window_min_ms: u64 = 2000,
     clock_synchronization_window_max_ms: u64 = 20000,
+    grid_iops_read_max: u64 = 16,
+    grid_iops_write_max: u64 = 16,
 };
 
 /// Configurations which are tunable per-cluster.