--- a/src/io/linux.zig
+++ b/src/io/linux.zig
@@ -16,10 +16,10 @@ pub const IO = struct {
 
     /// Operations not yet submitted to the kernel and waiting on available space in the
     /// submission queue.
-    unqueued: FIFO(Completion) = .{ .plot_id = .io_unqueued_count },
+    unqueued: FIFO(Completion) = .{ .name = "io_unqueued" },
 
     /// Completions that are ready to have their callbacks run.
-    completed: FIFO(Completion) = .{ .plot_id = .io_completed_count },
+    completed: FIFO(Completion) = .{ .name = "io_completed" },
 
     pub fn init(entries: u12, flags: u32) !IO {
         // Detect the linux version to ensure that we support all io_uring ops used.
@@ -107,7 +107,7 @@ pub const IO = struct {
         // Loop over a copy to avoid an infinite loop of `enqueue()` re-adding to `self.unqueued`.
         {
             var copy = self.unqueued;
-            self.unqueued = .{ .plot_id = copy.plot_id };
+            self.unqueued = .{ .name = copy.name };
             while (copy.pop()) |completion| self.enqueue(completion);
         }