--- a/src/lsm/grid.zig
+++ b/src/lsm/grid.zig
@@ -13,6 +13,7 @@ const SetAssociativeCache = @import("set_associative_cache.zig").SetAssociativeC
 const stdx = @import("../stdx.zig");
 
 const log = std.log.scoped(.grid);
+const tracer = @import("../tracer.zig");
 
 /// A block's type is implicitly determined by how its address is stored (e.g. in the index block).
 /// BlockType is an additional check that a block has the expected type on read.
@@ -56,9 +57,8 @@ pub fn GridType(comptime Storage: type) type {
     return struct {
         const Grid = @This();
 
-        // TODO put more thought into how low/high these limits should be.
-        pub const read_iops_max = 16;
-        pub const write_iops_max = 16;
+        pub const read_iops_max = constants.grid_iops_read_max;
+        pub const write_iops_max = constants.grid_iops_write_max;
 
         pub const BlockPtr = *align(constants.sector_size) [block_size]u8;
         pub const BlockPtrConst = *align(constants.sector_size) const [block_size]u8;
@@ -144,11 +144,13 @@ pub fn GridType(comptime Storage: type) type {
         cache: Cache,
 
         write_iops: IOPS(WriteIOP, write_iops_max) = .{},
+        write_iop_tracer_slots: [write_iops_max]?tracer.SpanStart = .{null} ** write_iops_max,
         write_queue: FIFO(Write) = .{},
 
         // Each read_iops has a corresponding block.
         read_iop_blocks: [read_iops_max]BlockPtr,
         read_iops: IOPS(ReadIOP, read_iops_max) = .{},
+        read_iop_tracer_slots: [read_iops_max]?tracer.SpanStart = .{null} ** read_iops_max,
         read_queue: FIFO(Read) = .{},
 
         // List if Read.pending's which are in `read_queue` but also waiting for a free `read_iops`.
@@ -320,6 +322,14 @@ pub fn GridType(comptime Storage: type) type {
         }
 
         fn write_block_with(grid: *Grid, iop: *WriteIOP, write: *Write) void {
+            const write_iop_index = grid.write_iops.index(iop);
+            tracer.start(
+                &grid.write_iop_tracer_slots[write_iop_index],
+                .{ .grid_write_iop = .{ .index = write_iop_index } },
+                .{ .grid_write_iop = .{ .index = write_iop_index } },
+                @src(),
+            );
+
             iop.* = .{
                 .grid = grid,
                 .completion = undefined,
@@ -351,6 +361,13 @@ pub fn GridType(comptime Storage: type) type {
             const cache_block = &grid.cache_blocks[cache_index];
             std.mem.swap(BlockPtr, cache_block, completed_write.block);
             std.mem.set(u8, completed_write.block.*, 0);
+
+            const write_iop_index = grid.write_iops.index(iop);
+            tracer.end(
+                &grid.write_iop_tracer_slots[write_iop_index],
+                .{ .grid_write_iop = .{ .index = write_iop_index } },
+                .{ .grid_write_iop = .{ .index = write_iop_index } },
+            );
 
             // Start a queued write if possible *before* calling the completed
             // write's callback. This ensures that if the callback calls
@@ -445,6 +462,14 @@ pub fn GridType(comptime Storage: type) type {
             // We can only update the cache if the Grid is not resolving callbacks with a cache block.
             assert(!grid.read_resolving);
 
+            const read_iop_index = grid.read_iops.index(iop);
+            tracer.start(
+                &grid.read_iop_tracer_slots[read_iop_index],
+                .{ .grid_read_iop = .{ .index = read_iop_index } },
+                .{ .grid_read_iop = .{ .index = read_iop_index } },
+                @src(),
+            );
+
             iop.* = .{
                 .completion = undefined,
                 .read = read,
@@ -471,6 +496,13 @@ pub fn GridType(comptime Storage: type) type {
             const cache_block = &grid.cache_blocks[cache_index];
             std.mem.swap(BlockPtr, iop_block, cache_block);
             std.mem.set(u8, iop_block.*, 0);
+
+            const read_iop_index = grid.read_iops.index(iop);
+            tracer.end(
+                &grid.read_iop_tracer_slots[read_iop_index],
+                .{ .grid_read_iop = .{ .index = read_iop_index } },
+                .{ .grid_read_iop = .{ .index = read_iop_index } },
+            );
 
             // Handoff the iop to a pending read or release it before resolving the callbacks below.
             if (grid.read_pending_queue.pop()) |pending| {