New traces:

* Trace start/end for every grid read_iop and write_iop.
* Trace start/end for every io flush and callback.
* Plot depth of every FIFO.
* Plot hit/miss count for every cache.
* Plot hit/miss count for filter blocks.

This surfaced https://github.com/tigerbeetledb/tigerbeetle/issues/551 and https://github.com/tigerbeetledb/tigerbeetle/issues/550 already.

Plus some internal changes:

* Upgrade tracy to v0.9.1. Mostly for https://github.com/wolfpld/tracy/issues/286.
* Remove support for perfetto, since noone seems to be using it (I included it originally as a low-setup option vs tracy).
* Intern strings in the tracy backend. Tracy requires certain strings to be unique. Previously I ensured that by requiring them to all be comptime constants, but this became pretty painful over time. Interning allows the api to be much simpler. This does require dynamic allocation, but only when the tracy backend is enabled. Nor  does it noticeably degrade performance when benchmarking with tracy enabled.

## Pre-merge checklist

Performance:

* [x] Compare `zig benchmark` on linux before and after this pr.
``` sh
#before
1223 batches in 119.52 s
load offered = 1000000 tx/s
load accepted = 83667 tx/s
batch latency p00 = 5 ms
batch latency p10 = 32 ms
batch latency p20 = 36 ms
batch latency p30 = 37 ms
batch latency p40 = 37 ms
batch latency p50 = 37 ms
batch latency p60 = 42 ms
batch latency p70 = 47 ms
batch latency p80 = 47 ms
batch latency p90 = 52 ms
batch latency p100 = 4366 ms
transfer latency p00 = 5 ms
transfer latency p10 = 3591 ms
transfer latency p20 = 10782 ms
transfer latency p30 = 18704 ms
transfer latency p40 = 28121 ms
transfer latency p50 = 39882 ms
transfer latency p60 = 53852 ms
transfer latency p70 = 66942 ms
transfer latency p80 = 81368 ms
transfer latency p90 = 95904 ms
transfer latency p100 = 109489 ms

# after
1223 batches in 125.95 s
load offered = 1000000 tx/s
load accepted = 79393 tx/s
batch latency p00 = 5 ms
batch latency p10 = 37 ms
batch latency p20 = 41 ms
batch latency p30 = 42 ms
batch latency p40 = 42 ms
batch latency p50 = 42 ms
batch latency p60 = 47 ms
batch latency p70 = 52 ms
batch latency p80 = 57 ms
batch latency p90 = 57 ms
batch latency p100 = 4019 ms
transfer latency p00 = 5 ms
transfer latency p10 = 3756 ms
transfer latency p20 = 11513 ms
transfer latency p30 = 20129 ms
transfer latency p40 = 30017 ms
transfer latency p50 = 42646 ms
transfer latency p60 = 57246 ms
transfer latency p70 = 71302 ms
transfer latency p80 = 86682 ms
transfer latency p90 = 101707 ms
transfer latency p100 = 115920 ms
```